Update
Build AI manager package
