+++ src/main/webapp/jsp/base/admin/manager.jsp
@@ -46,7 +46,7 @@
     <ul class="main-menu visible-on-click" id="aiChat">
       <li style="background-color:yellow;">AI chat</li>
       <li>
-        <a href="/aiChatManager/userList" target="_blank">user list</a>
+        <a href="/aiManager/userList" target="_blank">user list</a>
       </li>
     </ul><!-- aiChat -->
